Rename file to initializeGame
