+++ src/index.js
@@ -1,4 +1,4 @@
 import './stylesheets/index.scss';
-import initializeGame from './views/playerForm';
+import initializeGame from './views/initializeGame';
 
 window.addEventListener('DOMContentLoaded', initializeGame);
