+++ src/views/initializeGame.js
